feat(grid): implement data-driven spawning and new grid utilities
Introduces a data-driven spawning system with fallback support, replacing hardcoded entity spawning to improve flexibility and maintainability. Adds robust error handling and compatibility with existing components.

Enhances grid functionality with new constructors, utilities for row and column access, neighbor calculations, and position iterators. Refactors grid dimensions to use `u32` for consistency.

Includes updates to the map generation process, entity spawning logic, and integration of RON-based entity definitions. Adds comprehensive tests to ensure component compatibility and integration accuracy.

+++ src/model/model_constants.rs
@@ -1,6 +1,6 @@
 pub struct ModelConstants;
 
 impl ModelConstants {
-    pub const MAP_WIDTH: usize = 50;
-    pub const MAP_HEIGHT: usize = 24;
+    pub const MAP_WIDTH: u32 = 50;
+    pub const MAP_HEIGHT: u32 = 24;
 }
